v2.5.1.8 -> v2.5.1.9
- Russell King: large ARM update
- Adam Richter et al: more kdev_t updates

+++ include/asm-arm/hdreg.h
@@ -7,7 +7,7 @@
 #ifndef __ASMARM_HDREG_H
 #define __ASMARM_HDREG_H
 
-typedef unsigned long ide_ioreg_t;
+typedef unsigned int ide_ioreg_t;
 
 #endif /* __ASMARM_HDREG_H */
